Server update-> Migration from java 7 to java 8. Libraries updated. Objectify replaced with direct DataStore management.

+++ debugbackend/src/main/java/com/lucevent/newsup/debugbackend/net/Net.java
@@ -18,16 +18,16 @@ public class Net {
 	}
 
 	public static void sendStatus(int code, String name, int info,
-	                              long check_time, long check_duration, int check_rounds,
-	                              int num_news, int with_content)
+	                              long check_time, long check_duration, long check_rounds,
+	                              long num_news, long with_content)
 	{
 		try {
 			Connection.Response response = Jsoup.connect("http://newsup-2406.appspot.com/dev?new_status")
 					.method(Connection.Method.POST)
 					.postDataCharset("UTF-8")
 					.data("code", Integer.toString(code), "name", name, "info", Integer.toString(info),
-							"check_time", Long.toString(check_time), "check_duration", Long.toString(check_duration), "check_rounds", Integer.toString(check_rounds),
-							"num_news", Integer.toString(num_news), "no_content", Integer.toString(with_content)
+							"check_time", Long.toString(check_time), "check_duration", Long.toString(check_duration), "check_rounds", Long.toString(check_rounds),
+							"num_news", Long.toString(num_news), "no_content", Long.toString(with_content)
 					)
 					.execute();
 		} catch (Exception ignored) {
